refactor: remove ingest versions
