+++ indykite_sdk_node/src/sdk/__test__/ingest/new_client.test.ts
@@ -1,6 +1,6 @@
 import { IngestAPIClient } from '../../../grpc/indykite/ingest/v1beta2/ingest_api.grpc-client';
 import { SdkClient } from '../../client/client';
-import { IngestV2Client } from '../../ingest_v2';
+import { IngestClient } from '../../ingest';
 import { applicationTokenMock } from '../../utils/test_utils';
 
 afterEach(() => {
@@ -9,14 +9,14 @@ afterEach(() => {
 
 describe('when a new client is created', () => {
   describe('when no error is thrown', () => {
-    let returnedValue: IngestV2Client;
+    let returnedValue: IngestClient;
 
     beforeEach(async () => {
       jest
         .spyOn(SdkClient, 'createIdentityInstance')
         .mockImplementation(() => Promise.resolve({ client: {} } as SdkClient));
 
-      returnedValue = await IngestV2Client.createInstance(JSON.stringify(applicationTokenMock));
+      returnedValue = await IngestClient.createInstance(JSON.stringify(applicationTokenMock));
     });
 
     it('creates a new instance', () => {
@@ -24,7 +24,7 @@ describe('when a new client is created', () => {
         IngestAPIClient,
         JSON.stringify(applicationTokenMock),
       );
-      expect(returnedValue).toBeInstanceOf(IngestV2Client);
+      expect(returnedValue).toBeInstanceOf(IngestClient);
     });
   });
 
@@ -37,7 +37,7 @@ describe('when a new client is created', () => {
         .spyOn(SdkClient, 'createIdentityInstance')
         .mockImplementation(() => Promise.reject(error));
       try {
-        await IngestV2Client.createInstance(JSON.stringify(applicationTokenMock));
+        await IngestClient.createInstance(JSON.stringify(applicationTokenMock));
       } catch (err) {
         caughtError = err as Error;
       }
